requirements.txt
cycler==0.10.0
future==0.17.1
fysom==2.1.5
h5py==2.9.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.0rc2
opencv-python==4.1.0.25
pkg-resources==0.0.0
PyAudio==0.2.11
pygame==1.9.6
PyOpenGL==3.1.0
pyparsing==2.4.0
pyserial==3.4
python-dateutil==2.8.0
six==1.12.0
